created first call structure
