+++ OIIInspector/OIIIClient.py
@@ -1,0 +1,9 @@
+import json
+
+
+class OIIIClient:
+
+    def get_index_image_package_list(self, image_address, address_type):
+
+        print(image_address, address_type)
+        return json.loads("{}")
